Adds LCD_V15 code

+++ Arduino/Ardruino_timer/Timer_LCD/timer_LCD_V20/KWH_Grid.ino
@@ -3,11 +3,11 @@ Functions to calculate KWH and Grite time
 All calculations of time in milisecomds
 ********************************************************/
 
-float Cal_Gridtime(float amps, unsigned long previousMillis)
+unsigned long Cal_Gridtime(float amps, unsigned long previousMillis)
 {
   if (amps < 0)
   {
-    float gtime = millis() - previousMillis;
+    unsigned int gtime = millis() - previousMillis;
     return gtime;
   }
   return 0;
